+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -566,6 +566,13 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     }
 
     /**
+     * @return the first element which the [selector] yields the smallest value or null if empty
+     */
+    public inline fun <R : Comparable<R>> minByOrNull(selector: (element: T) -> R): T? {
+        return if (isEmpty()) null else minBy(selector)
+    }
+
+    /**
      * @return the first element which the [selector] yields the largest value.
      * @throws NoSuchElementException if this ImmutableArray is empty
      */