+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableBooleanArray.kt
@@ -563,6 +563,13 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
     }
 
     /**
+     * @return the first element which the [selector] yields the smallest value or null if empty
+     */
+    public inline fun <R : Comparable<R>> minByOrNull(selector: (element: Boolean) -> R): Boolean? {
+        return if (isEmpty()) null else minBy(selector)
+    }
+
+    /**
      * @return the first element which the [selector] yields the largest value.
      * @throws NoSuchElementException if this ImmutableBooleanArray is empty
      */